Rename from setup-storage-controller to snippt-list-controller
